fix: minor bugs

+++ src/yield-exposed-tokens/GenericYeToken.sol
@@ -19,7 +19,7 @@ contract GenericYeToken is YieldExposedToken {
         string calldata name_,
         string calldata symbol_,
         address underlyingToken_,
-        uint8 minimumReservePercentage_,
+        uint256 minimumReservePercentage_,
         address yieldVault_,
         address yieldRecipient_,
         address lxlyBridge_,
